requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
DateTime==5.4
Flask==3.0.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
gspread==6.0.2
gunicorn==21.2.0
httplib2==0.22.0
line-bot-sdk==3.7.0
nba_api==1.4.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
python-dotenv==1.0.1
platformdirs==4.2.0
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.2.0
Werkzeug==3.0.1